Fix devcontainer
The Haskell chain needs to run as the current user, I did not manage to
create a devcontainer where everything is owned by root. It failed
because HLS wants to run ghcup, so it would just install everything
under /home/vscode even though it was already installed under /opt.

It is probably possible to fix, but I did not want to spend more time on
it.